ambil jere

+++ database/migrations/2025_04_09_191537_create_destinations_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->text('gambar4');
             $table->text('gambar5');
             $table->text('gambarM');
-            $table->unsignedInteger('kategori_id');
+            $table->unsignedBigInteger('kategori_id');
             $table->text('desk');
             $table->text('long_desk');
             $table->integer('click_count')->nullable()->default(0);
